# openapi.yaml
openapi: 3.0.0
info:
  title: Snake Game API
  description: API for Snake Game with Auth, Leaderboard, and Spectator features.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - createdAt

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    SignupRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - username
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required:
        - user
        - token

    Position:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
      required:
        - x
        - y

    Direction:
      type: string
      enum: [UP, DOWN, LEFT, RIGHT]

    GameSession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
        isActive:
          type: boolean
        snake:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        food:
          $ref: '#/components/schemas/Position'
        direction:
          $ref: '#/components/schemas/Direction'
        startedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - score
        - isActive
        - snake
        - food
        - direction
        - startedAt

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
        createdAt:
          type: string
          format: date-time
        rank:
          type: integer
      required:
        - id
        - userId
        - username
        - score
        - createdAt
        - rank

    CreateSessionRequest:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
      required:
        - userId
        - username

    UpdateSessionRequest:
      type: object
      description: Partial updates for game session
      properties:
        score:
          type: integer
        snake:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        food:
          $ref: '#/components/schemas/Position'
        direction:
          $ref: '#/components/schemas/Direction'
        isActive:
          type: boolean

    EndSessionRequest:
      type: object
      properties:
        finalScore:
          type: integer
      required:
        - finalScore

    SubmitScoreRequest:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
      required:
        - userId
        - username
        - score

paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /auth/signup:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Username or email already exists

  /auth/logout:
    post:
      summary: Logout user
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully logged out

  /auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid or missing token

  /leaderboard:
    get:
      summary: Get leaderboard
      tags: [Leaderboard]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of top scores to return
      responses:
        '200':
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
    post:
      summary: Submit a score to the leaderboard
      tags: [Leaderboard]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitScoreRequest'
      responses:
        '201':
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardEntry'

  /sessions:
    get:
      summary: Get active game sessions
      tags: [Game Sessions]
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSession'
    post:
      summary: Create a new game session
      tags: [Game Sessions]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'

  /sessions/{sessionId}:
    get:
      summary: Get a specific session
      tags: [Game Sessions]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        '404':
          description: Session not found

    patch:
      summary: Update a session (e.g. movement, score)
      tags: [Game Sessions]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        '404':
          description: Session not found

  /sessions/{sessionId}/end:
    post:
      summary: End a game session
      tags: [Game Sessions]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndSessionRequest'
      responses:
        '200':
          description: Session ended
        '404':
          description: Session not found